add packagecloudext package for helper functions that use the API
This adds the `packagecloudext` package which includes helper functions based on functionality within the main `packagecloud` package. The first function is `NextDebRelease` which looks at the versions for a specific package and determines which release number should be next.

For example, assume you have package `testPackage` with the following versions:

```
0.0.1-1
0.0.2-1
0.0.2-2
```

If you were to get the next release number for `0.0.1`, it would be `2`. If you wanted the next release number for `0.0.2`, it would be `3`. Lastly, if you wanted the next release number for `0.0.3`, it would be `1`.